requirements.txt
mcp==1.27.0
pydantic==2.12.5
PyJWT==2.12.1
python-dotenv==1.2.2
